@@ -10,8 +10,9 @@
   "private": true,
   "dependencies": {
     "core-js": "^3.6.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.2.2",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.0.0"
   },
@@ -28,8 +29,8 @@
     "@testing-library/react-hooks": "7.0.2",
     "@types/jest": "27.0.2",
     "@types/node": "16.11.7",
-    "@types/react": "17.0.40",
-    "@types/react-dom": "17.0.13",
+    "@types/react": "latest",
+    "@types/react-dom": "latest",
     "@typescript-eslint/eslint-plugin": "~5.10.0",
     "@typescript-eslint/parser": "~5.10.0",
     "babel-jest": "27.2.3",
