@@ -1,11 +1,20 @@
 import { StrictMode } from 'react';
-import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import App from './app/App';
 
-import App from './app/app';
+const appContainer = document.getElementById('root');
 
-ReactDOM.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-  document.getElementById('root')
-);
+if (appContainer) {
+  const root = createRoot(appContainer);
+
+  root.render(
+    <StrictMode>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </StrictMode>
+  );
+} else {
+  console.error('Cannot find element with id "root"');
+}
